fixed formatting + added backbone test
