+++ keras_nlp/models/gpt_neox/gpt_neox_backbone.py
@@ -1,18 +1,33 @@
+# Copyright 2023 The KerasNLP Authors
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     https://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
 import tensorflow as tf
 from tensorflow import keras
-
-from keras_nlp.models.backbone import Backbone
-from keras_nlp.layers.position_embedding import PositionEmbedding
 from tensorflow.experimental import dtensor
 from tensorflow.experimental.dtensor import Layout
 
-from gpt_neox_decoder import GPTNeoXDecoder
+from keras_nlp.api_export import keras_nlp_export
+from keras_nlp.layers.position_embedding import PositionEmbedding
+from keras_nlp.models.backbone import Backbone
+from keras_nlp.models.gpt_neox.gpt_neox_decoder import GPTNeoXDecoder
+
 
 def _gpt_2_kernel_initializer(stddev=0.02):
     return keras.initializers.RandomNormal(stddev=stddev)
 
-class GPT2Backbone(Backbone):
 
+@keras_nlp_export("keras_nlp.models.GPT2Backbone")
+class GPTNeoXBackbone(Backbone):
     def __init__(
         self,
         vocabulary_size,
@@ -150,4 +165,4 @@ class GPT2Backbone(Backbone):
         layout_map[r".*_feedforward_output_dense.bias"] = Layout(
             [unshard_dim], mesh
         )
-        return layout_map   
+        return layout_map
